Move settings out of __init__.py
