--- esgf_core_utils/settings/kafka/__init__.py
@@ -1,5 +1,0 @@
-from esgf_core_utils.settings.kafka.consumer import ConsumerSettings
-from esgf_core_utils.settings.kafka.producer import ProducerSettings
-
-consumer_settings = ConsumerSettings()
-producer_settings = ProducerSettings()
