+++ esgf_core_utils/settings/kafka/settings.py
@@ -1,0 +1,5 @@
+from esgf_core_utils.settings.kafka.consumer import ConsumerSettings
+from esgf_core_utils.settings.kafka.producer import ProducerSettings
+
+consumer_settings = ConsumerSettings()
+producer_settings = ProducerSettings()
